pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Django backend for location-based search"
authors = ["MichellellehciM"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.6"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
